+++ tests/testgdi.c
@@ -57,13 +57,13 @@ win_draw(win_t *win)
 		int g = 0;
 		int b = 0;
 		
-		GdipCreatePen1 (a << 24 | r << 16 | g << 8 | b,
+		GdipCreatePen1 ((guint32)a << 24 | r << 16 | g << 8 | b,
 				10, UnitPixel, &pen);
 		
 		GdipDrawRectangle (gp, pen, 10, 10, 60, 60);
 		GdipDrawLine (gp, pen, 0, 0, 100, 100);
 		
-		GdipCreateSolidFill (a << 24 | r << 16 | g << 8 | b, &brush);
+		GdipCreateSolidFill ((guint32)a << 24 | r << 16 | g << 8 | b, &brush);
 		
 		printf ("%d\n",GdipFillEllipse (gp, (GpBrush*)brush, 40, 40, 50, 75));
 //		return;
